tests/inputdevice-mouse.html

<!DOCTYPE html>
<title>InputDevice test for mouse-triggered events</title>
<meta name="viewport" content="initial-scale=1">
<meta charset="UTF-8">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevice-tests.js"></script>
<style>
#target {
  border: 1px solid black;
  width: 100px;
  height: 100px;
  background-color: lightgreen;
  margin-bottom: 20px;
}
</style>

<h1>InputDevice - test for mouse-triggered events</h1>
<p>Double-click with a pointing device that doesn't fire touch events (eg. a mouse) in the box below, then click outside the box.</p>
<div id="target" tabindex=1></div>
<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceValidator(sourceDevice) {
  assert_true(sourceDevice != null, "sourceDevice is set");
  assert_false(sourceDevice.firesTouchEvents, "sourcedevice.firesTouchEvents");
}

function runTests() {
  var events = ["mousedown", "mousemove", "mouseup", "mouseenter", "mouseleave", "mouseover", "mouseout", "click", "dblclick", "focus", "blur"];

  // Mozilla doesn't support focusin/focusout: https://bugzilla.mozilla.org/show_bug.cgi?id=687787
  // There doesn't appear to be any good way to feature-detect for it.
  if (navigator.userAgent.indexOf("Firefox") == -1)
    events = events.concat(["focusin", "focusout"]);

  if ('PointerEvent' in window)
    events = events.concat(["pointerdown", "pointermove", "pointerup", "pointerenter", "pointerleave", "pointerover", "pointerout"]);

  runInputDeviceEventTests(events, inputDeviceValidator);
}
</script>
